Remove all unnecessary static_cast.
Enums are now not scoped anymore to allow for implicit int conversion.
This is a bit of a bummer, but I will make a point to use them as
scoped, and hopefully all (or most) unscoped usages have been removed. I
still think this is worth  it, as the static_cast's got to be a bit
much.

+++ tests/castlingtypetest.cpp
@@ -6,6 +6,6 @@ using namespace goldfish;
 
 TEST(castlingtypetest, test_values) {
     for (auto castlingtype : CastlingTypes::values) {
-        EXPECT_EQ(castlingtype, CastlingTypes::values[static_cast<int>(castlingtype)]);
+        EXPECT_EQ(castlingtype, CastlingTypes::values[castlingtype]);
     }
 }
