+++ tests/filetest.cpp
@@ -6,6 +6,6 @@ using namespace goldfish;
 
 TEST(filetest, test_values ) {
     for (auto file : Files::values) {
-        EXPECT_EQ(file, Files::values[static_cast<int>(file)]);
+        EXPECT_EQ(file, Files::values[file]);
     }
 }
