+++ tests/ranktest.cpp
@@ -6,6 +6,6 @@ using namespace goldfish;
 
 TEST(ranktest, test_values) {
     for (auto rank : Ranks::values) {
-        EXPECT_EQ(rank, Ranks::values[static_cast<int>(rank)]);
+        EXPECT_EQ(rank, Ranks::values[rank]);
     }
 }
